Add CI mypy/coverage gate, type fixes, and site landing updates.
Enforce static typing and a 78% coverage floor in CI, resolve mypy
findings across the engine, and refresh the marketing site.

+++ src/driftless/judges.py
@@ -136,7 +136,7 @@ def _pearson(xs: list[float], ys: list[float]) -> float | None:
     dy = sum((y - my) ** 2 for y in ys) ** 0.5
     if dx == 0 or dy == 0:
         return None
-    return num / (dx * dy)
+    return float(num / (dx * dy))
 
 
 def judge_agreement(
